♻️ Rename Models directory to a domain related directory name
